:sparkles: Split fields by function
Before this commit, all fields were added to just one list. However, due
to addin support to composed PKs, we need fields for schema, fields to
be filtered, columns for the database.

So, a bunch of lists with different goal were added to the
`models.Model`. This way, templates can handle it properly.