:sparkles: Generator (#13)
* :sparkles: Implement generating schemas

* :sparkles: Implement generating query, resultset and store

* :bug: Predicts the existing of composed interfaces

Composed interfaces were not predicted what led the lib to error when
finding them. This commit do not proper handle them, but gracefully
ignores them.

* :sparkles: Split fields by function

Before this commit, all fields were added to just one list. However, due
to addin support to composed PKs, we need fields for schema, fields to
be filtered, columns for the database.

So, a bunch of lists with different goal were added to the
`models.Model`. This way, templates can handle it properly.

* :sparkles: Add composite PKs on the generator